pineapple-app/server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pillow==10.4.0
ultralytics>=8.3.50
